+++ editor/include/console_view.h
@@ -10,6 +10,11 @@
 #include <vector>
 
 #include "include/glview.h"
+#include "include/glview_listener.h"
+#include "include/console_view.h"
+#include "include/save_button.h"
+#include "include/switcher.h"
+#include "include/backtrace.h"
 
 class ConsoleView : public GLView {
  public:
@@ -53,6 +58,23 @@ class ConsoleView : public GLView {
   bool content_changed_;
   // If content not changes, display by timer.
   timeval last_display_;
+};
+
+class SaveButton;
+class ConsoleViewListener : public GLViewListener {
+ public:
+  ConsoleViewListener(ConsoleView* console_view, SaveButton* save_button,
+                      Switcher* draw_grid_switcher,
+                      Switcher* colorizer_switcher, Backtrace* backtrace);
+
+  virtual void KeyPressed(uint8_t key, int x, int y);
+
+ private:
+  ConsoleView* console_view_;
+  SaveButton* save_button_;
+  Switcher* draw_grid_switcher_;
+  Switcher* colorizer_switcher_;
+  Backtrace* backtrace_;
 };
 
 #endif  // EDITOR_INCLUDE_CONSOLE_VIEW_H_
